+++ src/commonMain/kotlin/csense/kotlin/tests/assertions/String.kt
@@ -2,7 +2,6 @@
 
 package csense.kotlin.tests.assertions
 
-import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotEquals
 import kotlin.test.assertTrue