+++ changelog.md
@@ -1,3 +1,6 @@
+# 0.0.47
+- added assertCallbackCalledWith 
+
 # 0.0.46
 
 - more Extensions